more api cleanups
