+++ types/src/boolean/mod.rs
@@ -42,9 +42,10 @@ mod boolean;
 pub use self::boolean::*;
 
 pub mod prelude {
-    //! Includes non-mapping types for the `boolean` type.
+    //! Includes all types for the `boolean` type.
     //!
     //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
 
     pub use super::boolean::*;
+    pub use super::mapping::*;
 }
