+++ types/src/date/mod.rs
@@ -111,7 +111,7 @@ pub type ChronoDateTime = DateTime<UTC>;
 pub type DefaultDateFormat = BasicDateTime;
 
 pub mod prelude {
-    //! Includes non-mapping types for the `date` type.
+    //! Includes all types for the `date` type.
     //!
     //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
 
@@ -119,4 +119,5 @@ pub mod prelude {
     pub use super::format::*;
     pub use super::date::*;
     pub use super::formats::*;
+    pub use super::mapping::*;
 }
