+++ types/src/geo/mod.rs
@@ -70,7 +70,7 @@ pub mod mapping {
 }
 
 pub mod prelude {
-    //! Includes non-mapping types for the `geo_point` and `geo_shape` types.
+    //! Includes all types for the `geo_point` and `geo_shape` types.
     //!
     //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
