+++ types/src/geo/point/mod.rs
@@ -59,7 +59,7 @@ pub use self::formats::*;
 pub type DefaultGeoPointFormat = GeoPointArray;
 
 pub mod prelude {
-    //! Includes non-mapping types for the `geo_point` type.
+    //! Includes all types for the `geo_point` type.
     //!
     //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
 
@@ -67,4 +67,5 @@ pub mod prelude {
     pub use super::format::*;
     pub use super::point::*;
     pub use super::formats::*;
+    pub use super::mapping::*;
 }
