+++ types/src/geo/shape/mod.rs
@@ -49,9 +49,10 @@ mod shape;
 pub use self::shape::*;
 
 pub mod prelude {
-    //! Includes non-mapping types for the `geo_shape` types.
+    //! Includes all types for the `geo_shape` types.
     //!
     //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
 
     pub use super::shape::*;
+    pub use super::mapping::*;
 }
