+++ types/src/number/mod.rs
@@ -56,9 +56,10 @@ mod number;
 pub use self::number::*;
 
 pub mod prelude {
-    //! Includes non-mapping types for the `number` type.
+    //! Includes all types for the `number` type.
     //!
     //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
 
     pub use super::number::*;
+    pub use super::mapping::*;
 }
