+++ types/src/string/keyword/keyword.rs
@@ -2,7 +2,7 @@ use std::marker::PhantomData;
 use serde::{Serialize, Deserialize, Serializer, Deserializer};
 use serde::de::{Visitor, Error};
 use super::mapping::{KeywordMapping, KeywordFormat};
-use ::mapping::FieldType;
+use ::field::FieldType;
 
 /// An Elasticsearch `keyword` with a mapping.
 ///
