+++ types/src/string/keyword/mod.rs
@@ -9,3 +9,12 @@ pub mod mapping;
 mod keyword;
 
 pub use self::keyword::*;
+
+pub mod prelude {
+	//! Includes all types for the `keyword` type.
+    //!
+    //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
+
+    pub use super::keyword::*;
+    pub use super::mapping::*;
+}
