+++ types/src/string/text/mod.rs
@@ -8,3 +8,12 @@ pub mod mapping;
 mod text;
 
 pub use self::text::*;
+
+pub mod prelude {
+	//! Includes all types for the `text` type.
+    //!
+    //! This is a convenience module to make it easy to build mappings for multiple types without too many `use` statements.
+
+    pub use super::text::*;
+    pub use super::mapping::*;
+}
