+++ types/src/string/text/text.rs
@@ -2,7 +2,7 @@ use std::marker::PhantomData;
 use serde::{Serialize, Deserialize, Serializer, Deserializer};
 use serde::de::{Visitor, Error};
 use super::mapping::{TextMapping, TextFormat};
-use ::mapping::FieldType;
+use ::field::FieldType;
 
 /// An Elasticsearch `text` field with a mapping.
 ///
